refactor : works 페이지 카테고리 리팩토링
